Initial commit of 'util_FindCode-Results-example.vbs'
The purpose of this macro is to demonstrate how to pop a particular Find form, select an item, and return one of its field values to a variable. This value is then used to locate the data and add it to the Results, which are displayed at the end of the process. This can be useful to show the user the records that were created, edited, or whatever so that they can review them if they like.