Introduce MacroProcess messages and results
Before this change, debugging messages used MsgBox (with some leftover calls to WScript.Echo) to present the user with information about what was happening as the script ran. Also, when the script finished, the user was not presented with any information about what records were added or edited.
    
After this change, debugging messages are logged using MacroProcess.AddMessage. This way, the information is available, but there is no need to respond to so many message boxes. Also, when the script finishes, the Employee records that were added/edited are presented to the user in the Results window.